pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.5",
    "boto3>=1.42.49",
    "cryptography>=46.0.5",
    "fastapi>=0.128.8",
    "psycopg[binary,pool]>=3.2.13",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.39.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
